refactor(cli): dedup watch config pipeline and simplify shutdown
- Extract the shared config pipeline (migration + file config + CLI options
  + merge) into buildMergedConfig in defaultAction, used by both the default
  and watch actions instead of being duplicated.
- Extract the inline --watch conflict checks in cliRun into a named
  validateWatchOptions function.
- Simplify watch shutdown: create the keep-alive promise up front so cleanup
  always resolves it (removing the cleanupResolve/cleanupDone race dance and
  the SIGINT/SIGTERM wrapper functions), while keeping idempotency and the
  in-flight-rebuild await.
- Move buildWatchIgnoreFilter into its own watchIgnore module, bringing
  watchAction.ts under the 250-line guideline.
